Aegis/backend/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
bcrypt==4.1.2
passlib==1.7.4
pyjwt==2.8.0
httpx==0.26.0
pydantic==2.5.3